Re-seed image if needed
If the total loss from previous iteration is 0, the image is re-seeded, otherwise a "function value changing less than tolX" message will appear and fail.

Tested with many images and has worked every time. Note that you are more likely to reproduce the issue the larger the image output size is. Try 1024-1568.